refactor(api): split completion streaming from activity save logic
Replace completeActivityStep with streamCompletionEventsStep for the
already-completed case. No DB writes needed — just stream the SSE event
so the UI can detect completion and redirect. Save steps handle their
own completion during normal generation flow.
